services/rating-engine/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
structlog==24.4.0
python-dotenv==1.0.1
